remove unnecessary workflow
