+++ tests/exporters/common/test_tasks_manager.py
@@ -177,3 +177,13 @@ class TasksManagerTestCase(TestCase):
 
         model = TasksManager.get_model_from_task("question-answering", "uclanlp/visualbert-vqa")
         self.assertTrue(isinstance(model, VisualBertForQuestionAnswering))
+
+    def test_library_detection(self):
+        self.assertEqual(
+            TasksManager.infer_library_from_model("intfloat/multilingual-e5-large"), "sentence_transformers"
+        )
+        self.assertEqual(
+            TasksManager.infer_library_from_model("stabilityai/stable-diffusion-xl-base-1.0"), "diffusers"
+        )
+        self.assertEqual(TasksManager.infer_library_from_model("gpt2"), "transformers")
+        self.assertEqual(TasksManager.infer_library_from_model("timm/mobilenetv3_large_100.ra_in1k"), "timm")
