--- tests/exporters/test_tasks_manager.py
@@ -1,30 +1,0 @@
-# coding=utf-8
-# Copyright 2022 The HuggingFace Team. All rights reserved.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-import unittest
-
-from optimum.exporters import TasksManager
-
-
-class TestTasksManager(unittest.TestCase):
-    def test_library_detection(self):
-        self.assertEqual(
-            TasksManager.infer_library_from_model("intfloat/multilingual-e5-large"), "sentence_transformers"
-        )
-        self.assertEqual(
-            TasksManager.infer_library_from_model("stabilityai/stable-diffusion-xl-base-1.0"), "diffusers"
-        )
-        self.assertEqual(TasksManager.infer_library_from_model("gpt2"), "transformers")
-        self.assertEqual(TasksManager.infer_library_from_model("timm/mobilenetv3_large_100.ra_in1k"), "timm")
